+++ views/balance-checker.pug
@@ -11,6 +11,7 @@ html
         script(src="/scripts/checker/checker-frontend.js", type="text/javascript", defer="defer")
         script(src="/scripts/checker/checker-search.js", type="text/javascript", defer="defer")
         script(src="/scripts/checker/checker-events.js", type="text/javascript", defer="defer")
+        script(src="/scripts/checker/checker-error-validation.js", type="text/javascript", defer="defer")
         script(src="/scripts/checker/checker-resources.js", type="text/javascript", defer="defer")
         script(src="/scripts/checker/checker-main.js", type="text/javascript", defer="defer")
